# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ads Marketplace API",
    "version": "1.0.0",
    "description": "API documentation for the Ads Marketplace platform - connecting advertisers with channel owners on Telegram",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authorization token"
      },
      "TelegramAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Telegram-Init-Data",
        "description": "Telegram Mini App init data"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "telegramId": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "languageCode": {
            "type": "string",
            "nullable": true
          },
          "isPremium": {
            "type": "boolean"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          },
          "isAdvertiser": {
            "type": "boolean"
          },
          "isChannelOwner": {
            "type": "boolean"
          },
          "walletAddress": {
            "type": "string",
            "nullable": true
          },
          "onboardingCompleted": {
            "type": "boolean"
          },
          "wallets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "address": {
                  "type": "string"
                },
                "isMain": {
                  "type": "boolean"
                },
                "label": {
                  "type": "string",
                  "nullable": true
                },
                "verified": {
                  "type": "boolean"
                }
              }
            }
          },
          "stats": {
            "type": "object",
            "properties": {
              "channels": {
                "type": "integer"
              },
              "dealsAsAdvertiser": {
                "type": "integer"
              },
              "dealsAsPublisher": {
                "type": "integer"
              },
              "briefs": {
                "type": "integer"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user profile",
        "description": "Returns the authenticated user's complete profile including wallets and statistics",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update current user profile",
        "description": "Updates the authenticated user's profile settings",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isAdvertiser": {
                    "type": "boolean",
                    "description": "Set user as advertiser"
                  },
                  "isChannelOwner": {
                    "type": "boolean",
                    "description": "Set user as channel owner"
                  },
                  "languageCode": {
                    "type": "string",
                    "description": "User's preferred language"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/complete-onboarding": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Complete user onboarding with role selection",
        "description": "Sets user roles and marks onboarding as completed",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "isAdvertiser",
                  "isChannelOwner"
                ],
                "properties": {
                  "isAdvertiser": {
                    "type": "boolean"
                  },
                  "isChannelOwner": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Onboarding completed successfully"
          }
        }
      }
    },
    "/api/users/telegram-auth/status": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Telegram account connection status",
        "description": "Returns MTProto user session status for the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Telegram auth status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "telegramAuth": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/telegram-auth/start": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Start Telegram account connection",
        "description": "Sends verification code via Telegram app or SMS",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phoneNumber"
                ],
                "properties": {
                  "phoneNumber": {
                    "type": "string"
                  },
                  "forceSms": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code request result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "telegramAuth": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or request failed"
          }
        }
      }
    },
    "/api/users/telegram-auth/code": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Submit Telegram verification code",
        "description": "Verifies code and may transition to password-required state",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "telegramAuth": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code"
          }
        }
      }
    },
    "/api/users/telegram-auth/password": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Submit Telegram 2FA password",
        "description": "Completes Telegram account connection when password is required",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password accepted, Telegram account connected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "telegramAuth": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid password or session state"
          }
        }
      }
    },
    "/api/users/telegram-auth/disconnect": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Disconnect Telegram account session",
        "description": "Removes stored MTProto session for the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Telegram account disconnected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Disconnect failed"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get public user profile",
        "description": "Returns public information about a user including reviews and ratings",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}/reviews": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user's reviews",
        "description": "Returns paginated list of reviews for a specific user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reviews": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}